add function_mean calculation to utils
